chore:a tagDto class representation of a tag in response.

+++ src/main/java/com/projects/blog_application/domain/dtos/TagResponseDTO.java
@@ -13,5 +13,5 @@ import java.util.UUID;
 public class TagResponseDTO {
     private UUID id;
     private String name;
-    private int postCount;
+    private Long postCount;
 }
